Added ability to change auto save interval from preferences form

+++ AutoSaveModule.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Threading;
 using System.IO;
+using System.Diagnostics;
 
 using ProjectMemo.ProjectMemoConsole;
 using ProjectMemo.ProjectMemoConsole.CommandAttributes;
@@ -16,48 +17,51 @@ namespace ProjectMemo
         private static AutoSaveModule ActiveInstance;
 
         private string mSaveDirectory;
-        private CustomControls.CustomRichTextBox mRtbSurrogate;
+        private bool mRunLoop = false;
+        private Stopwatch mStopWatch;
 
         public AutoSaveModule() {
             ActiveInstance = this;
             mSaveDirectory = Directory.GetCurrentDirectory() + "\\" + "autosaves\\";
-            mRtbSurrogate = new CustomControls.CustomRichTextBox();
+            mStopWatch = new Stopwatch();
         }
 
         public void Run() {
             CustomConsole.Log("Started AutoSaveModule", true);
+            mRunLoop = true;
+            mStopWatch.Start();
 
-            while (Thread.CurrentThread.ThreadState == ThreadState.Running) {
-                Console.WriteLine(ProjectMemo.Forms.MainForm.AutoSaveInterval);
+            while (mRunLoop) {
+                //Console.WriteLine(mStopWatch.ElapsedMilliseconds);
+                if (mStopWatch.ElapsedMilliseconds > ProjectMemo.Forms.MainForm.AutoSaveInterval) {
+                    Dictionary<string, string> rtbData = ProjectMemo.Forms.MainForm.RichTextBoxData;
+                    int saveId = 0;
 
-                Thread.Sleep(ProjectMemo.Forms.MainForm.AutoSaveInterval);
+                    foreach (KeyValuePair<string, string> pair in rtbData) {
+                        string[] fullPathSplit = pair.Key.Split('\\');
+                        string fileName = fullPathSplit[fullPathSplit.Length - 1];
 
-                Dictionary<string, string> rtbData = ProjectMemo.Forms.MainForm.RichTextBoxData;
-                int saveId = 0;
+                        string autoSaveFilePath = mSaveDirectory + string.Format("autosave_{0}_{1}.tmp", DateTime.Now.ToString("yyyyMMdd_hhmmss"), saveId);
 
-                foreach (KeyValuePair<string, string> pair in rtbData) {
-                    string[] fullPathSplit = pair.Key.Split('\\');
-                    string fileName = fullPathSplit[fullPathSplit.Length - 1];
+                        if (File.Exists(autoSaveFilePath)) {
+                            File.Delete(autoSaveFilePath);
+                        }
 
-                    string autoSaveFilePath = mSaveDirectory + string.Format("autosave_{0}_{1}.tmp", DateTime.Now.ToString("yyyyMMdd_hhmmss"), saveId);
-
-                    if (File.Exists(autoSaveFilePath)) {
-                        File.Delete(autoSaveFilePath);
+                        File.WriteAllText(autoSaveFilePath, pair.Value);
+                        CustomConsole.Log("Saved temp file to: " + autoSaveFilePath);
+                        saveId++;
                     }
 
-                    File.WriteAllText(autoSaveFilePath, pair.Value);
-                    CustomConsole.Log("Saved temp file to: " + autoSaveFilePath);
-                    saveId++;
+                    mStopWatch.Restart();
                 }
             }
 
             CustomConsole.Log("Closing the AutoSaveModule thread.");
+            mStopWatch.Stop();
         }
 
         private void Stop() {
-            try {
-                Thread.CurrentThread.Abort();
-            } catch { return; }
+            mRunLoop = false;
         }
 
         [CommandMethod("autosave.stop", "")]
@@ -67,7 +71,8 @@ namespace ProjectMemo
 
         [CommandMethod("autosave.interval", "")]
         public static void LogInternal(string[] aArgs) {
-            CustomConsole.Log("AutoSaveInterval: " + ProjectMemo.Forms.MainForm.AutoSaveInterval);
+            int mins = (ProjectMemo.Forms.MainForm.AutoSaveInterval / 1000) / 60;
+            CustomConsole.Log("AutoSaveInterval: " + mins + " minutes (" + ProjectMemo.Forms.MainForm.AutoSaveInterval + ")");
         }
 
         [CommandMethod("autosave.clearfiles", "")]
